Fix for VC6
git-svn-id: http://miranda.googlecode.com/svn/trunk@7416 83df23c2-51fd-11de-bb24-2d5e9b4af41d

+++ miranda/protocols/MSN/msn_svcs.cpp
@@ -663,7 +663,7 @@ static int MsnGetAvatarInfo(WPARAM wParam,LPARAM lParam)
 		strcpy( AI->filename + len, "*" );
 
 		_finddata_t c_file;
-		intptr_t hFile = _findfirst(AI->filename, &c_file);
+		long hFile = _findfirst(AI->filename, &c_file);
 
 		// Find first .c file in current directory 
 		if (hFile == -1L)
